requirements.txt
openai
streamlit
Pillow
google-generativeai
beautifulsoup4
requests
toml
numpy